Rename project and ensure geowave CSV ingest works
